testing rearrangements
