--- fasta/pFasta/vcfToPfa_test.go
@@ -1,12 +1,0 @@
-package pFasta
-
-// import (
-// 	"fmt"
-// 	"testing"
-// )
-
-// func TestVcfToPfa(t *testing.T) {
-// 	// 100 bp
-
-// 	fmt.Print("")
-// }
